refactor(src/types): zod form validation Base 타입들을 별도의 파일로 분리
- /src/types/FormValidationBaseTypes.ts 위치에 존재
